Renamed the folder for easier capatability with Python. Fixed the issues with accessing files from a folder.
