+++ UserInterface/AddInventoryObj.py
@@ -10,7 +10,7 @@
 
 from PyQt5 import QtCore, QtGui, QtWidgets
 
-import RetrieverEssentials_rc
+import UserInterface.RetrieverEssentials_rc
 
 
 class Ui_StaffAddObj(object):
